\name{elapsedtime}
\alias{elapsedtime}
\title{Elapsed Time}
\description{
Compute the elapsed time between start time and end time
}
\usage{
elapsedtime(st, et)
}
\arguments{
  \item{st}{starting time}
  \item{et}{ending time}
}
\details{
It returns the time the process took to run.
}
\value{
String of the form \code{hh:mm:ss}
}
\note{It is not intended to be called directly.
}
\author{Washington Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\seealso{\code{\link{mnimput}}}

\keyword{NA}
\keyword{multivariate}
\keyword{ts}
\keyword{smooth}
