\name{Internal}
\alias{em.arima}
\alias{em.spline}
\alias{em.gam}
\alias{em.arrangemat}
\alias{em.arrangevec}
\alias{em.contribt1}
\alias{em.contribt2}
\alias{em.correctmat}
\alias{em.countmvec}
\alias{em.countmnear}
\alias{em.det}
\alias{em.dispersion}
\alias{em.existna}
\alias{em.extractcoord}
\alias{em.mean}
\alias{em.multiplyvec}
\alias{em.partmu}
\alias{em.partsigma}
\alias{em.putvalue}
\alias{em.rearrangemat}
\alias{em.rearrangevec}
\alias{em.replacewmean}
\alias{em.splitvec}
\alias{em.trace}
\alias{em.filter}
\alias{em.recursion}
\alias{em.nofilter}

\title{Internal function
}
\description{
Internal functions called from within \code{\link{mnimput}}
}
\details{
These functions are not intendend for user-level call.
}
\author{Washington Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\keyword{internal}
