\name{miss}
\alias{miss}
\docType{data}
\title{Sample Dataset}
\description{
A small sample dataset for the tutorial on data imputation
}
\usage{
	data(miss)
}
\format{
  A data frame with 24 observations on the following 5 variables.
  \describe{
    \item{\code{c31}}{a numeric vector with 1 missing observation}
    \item{\code{c32}}{a numeric vector with 1 missing observation}
    \item{\code{c33}}{a numeric vector with 6 missing observations}
    \item{\code{c34}}{a numeric vector with 3 missing observations}
    \item{\code{c35}}{a numeric vector with 3 missing observations}
  }
}
\examples{
data(miss)
}
\keyword{datasets}
