\name{mkjnw}
\alias{mkjnw}
\title{Example from Johnson \& Wichern's Book
}
\description{
Create a data matrix from the Johnson \& Wichern's book
}
\usage{
mkjnw()
}
\details{
This function creates a data matrix from the Johnson & Wichern's book.
}
\value{
It returns a data matrix.
}
\references{
Johnson, R., Wichern, D. (1998) \emph{Applied Multivariate Statistical Analysis}. Prentice Hall.
}
\author{Washington Junger \email{wjunger@ims.uerj.br} and Antonio Ponce de Leon \email{ponce@ims.uerj.br}
}
\seealso{\code{\link{mnimput}}}
\examples{
d <- mkjnw()
}

\keyword{NA}
\keyword{multivariate}
\keyword{ts}
\keyword{smooth}
